+++ src/apis/cards/types.ts
@@ -1,5 +1,5 @@
-import isValidDate from '@/utils/isValidDate';
 import { z } from 'zod';
+import isValidDate from '@/utils/isValidDate';
 
 const IMAGE_URL = 'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/task_image/';
