+++ src/apis/members/queries.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { deleteMember, getMembers } from '.';
-import { DeleteMemberRequest, GetMembersRequest } from './types';
+import { deleteMember, getMembers } from '@/apis/members/';
+import { DeleteMemberRequest, GetMembersRequest } from '@/apis/members/types';
 
 export const useMembersQuery = (params: GetMembersRequest) => {
   return useQuery({
